Added CardExporter class.
Created a CardExporter class to instantiate canvases for card images to be drawn onto. The instantiation allows for a new canvas to be created once a canvas is full of images (as determined by number of images per card by the game manager). The cards are therefore differentiated and exported separate from each other.